added first version of arm adjustment within menu
on main menu display, 'arm' replaces 'awepipe.' 'awepipe' has been transferred to the position of 'nah' on the side menu, along with the program loop. added a 'off' after each program loop for the medium motor. display was made for arm adjustment, with a 'up,' 'down,' and 'back.' the whole arm function is set in a 10 loop so it can be cancelled by pressing the center button (this is how 'back' works). the arm is placed in an unlimited loop, so it only cancels w/ 'back.' there is a switch block that leads to either the up or down button, and the respective arm direction. pressing the top button sends the program into a loop of moving the arm up until the button is released. same concept for down. possible tweaks: idk whether positive motor movement goes to move the arm up or down, that is to be tested. the amount and speed of med motor movement could also be adjusted. this is being uploaded so i can download to robot via bluetooth to test since i cant do it w/ this computer.